Update JSON API server URL (bacon unofficial cm-13.0 -- new)
Due to recent updates in the telephony stack, the telephony database
needs to be manually deleted upon upgrade. We don't want to do this
every time, so the deletion will be automated as part of the first build
that includes this updated OTA channel. This will force everyone to go
through the deletion before moving on to newer builds.